added EEPROM read and write for remaining volume
EEPROM read and write for the remaining volume on each container was
added, this data will be communicated via a UART packet to the mobile
app.

EEPROM now changes after dispense is done.

+++ firmware/include/USART.h
@@ -6,5 +6,5 @@
 void USART_init(unsigned int ubrr);
 void USART_send_char(char ch);
 void USART_send_string(char *str);
-void sendScaleReading(double weight);
+void sendScaleReading(float weight);
 #endif // !DEBUG
